requirements.txt
# Core dependencies
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0.3
aiohttp>=3.8.5
cryptography>=41.0.3
pyyaml>=6.0.1
python-binance>=1.0.19
ccxt>=4.1.13
ta>=0.10.2
pytz>=2023.3
tzlocal>=5.0.1
schedule>=1.2.0
psutil>=5.9.5
colorama>=0.4.6
termcolor>=2.3.0
curses-menu>=0.9.0
python-telegram-bot>=20.6

# Machine Learning
shap>=0.42.1
tensorflow-macos
tensorflow-metal
tensorflow-probability
tensorflow-hub
tensorflow-datasets
tensorflow-estimator
tensorflow-io-gcs-filesystem
tensorflow-metadata
tensorflow-model-optimization
tensorflow-serving-api>=2.16.1
tensorflow-text
tensorflow-decision-forests>=1.8.0
gast
protobuf
typing-extensions

# Deep Learning
torch>=2.0.1
torchvision>=0.15.2
tensorboard>=2.13.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gym>=0.26.0

# Natural Language Processing
transformers>=4.30.0
tokenizers>=0.13.3

# Explainable AI
lime>=0.2.0.1
plotly>=5.18.0

# Testing and Development
pytest-cov>=4.1.0
hypothesis>=6.80.0
sphinx>=7.1.0
sphinx-rtd-theme>=1.2.0
nbsphinx>=0.9.0
pre-commit>=3.3.0
mypy>=1.3.0
isort>=5.12.0
flake8>=6.0.0

# Additional ML and integration packages
openai>=0.27.0
joblib>=1.2.0
optuna>=3.6.0
multiprocess>=0.70.14

# Logging and notification
colorlog>=6.8.0

# Sentiment analysis
newsapi-python>=0.2.7
nltk>=3.8.1
textblob>=0.17.1
tweepy>=4.14.0
vaderSentiment>=3.3.2

# Monte Carlo simulation
multiprocess>=0.70.14  # Already included above, kept for clarity